+++ src/common/constants/index.js
@@ -1,7 +1,15 @@
 import events from "./eventsLib";
 import actions from "./actions";
-
+const cssClasses = {
+    openFacet:"UNX-accordian-open",
+    closeFacet: "UNX-accordian-close",
+    openBtn:"UNX-facet-open",
+    closeBtn:"UNX-facet-close",
+    viewMore:"UNX-view-more",
+    viewLess:"UNX-view-less"
+}
 export {
     events,
-    actions
+    actions,
+    cssClasses
 };