+++ src/modules/pagination/setPagination.js
@@ -1,14 +1,20 @@
 
 import renderPagination from './renderPagination';
+import infiniteScroller from './infiniteScroller';
+import debounce from '../utils/debounce';
 import {
     renderNewResults,
     paginationAction
 } from './actions';
+const onInfinteScroll = function(){
+    debounce(infiniteScroller.bind(this),1000)();
+}
 const setPagination = (prototype) => {
     prototype = Object.assign(prototype,{
         renderPagination,
         renderNewResults,
-        paginationAction
+        paginationAction,
+        onInfinteScroll
     })
 }
 
@@ -16,5 +22,6 @@ export {
     setPagination as default,
     renderPagination,
     renderNewResults,
-    paginationAction
+    paginationAction,
+    onInfinteScroll
 };