+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 ## Changelog
 
-##### 0.0.1
 
-- 🎉 First Commit!
+
+## [0.1.9] - 2020-09-17
+
+### Changed
+- fixed selected facet click issue 
+
+
+
+
+## [0.1.8] - 2020-09-16
+
+### Changed
+- facet configuration changes, only need to add the facetsEl under facet
+- removed rangeFacetEl and multiLevelFacteEl, from facet config object
+- onCallBack named to on Event
+- no need of 'updateConfig' as mandatory step
+
+
+### Added
+- configuration for view more functionality added, updated in read me
+- added animaion support for the facets collapse
+- added support for positioning the facets